Port telescope commands to a sep file
