+++ src/plugin.rs
@@ -5,6 +5,7 @@ use crate::{action::Action, app::App};
 pub mod plugin_popup;
 pub mod sfs_telescope;
 pub mod telescope;
+pub mod telescope_commands;
 pub trait Plugin {
     fn display_details(&self) -> String {
         String::from("Not implemented")
